# frontend/contract/swagger.yaml
openapi: 3.1.0
info:
  title: Trailarr API
  summary: Trailarr API available commands.
  description: API for Trailarr application.
  termsOfService: https://github.com/nandyalu/trailarr
  contact:
    name: Trailarr - Github
    url: https://github.com/nandyalu/trailarr
  license:
    name: GNU GPL 3.0
    url: https://github.com/nandyalu/trailarr/blob/main/LICENSE
  version: 0.4.0
servers:
  - url: '{protocol}://{hostpath}'
    variables:
      protocol:
        enum:
          - http
          - https
        default: http
      hostpath:
        default: 127.0.0.1:7889
paths:
...
  /api/v1/files/files:
    get:
      tags:
        - Files
        - Media
      summary: Get Files
      description: "Get files in a directory.\n\nArgs:\n    path (str): Path to the\
        \ directory.\nReturns:\n    FolderInfo|None: Folder information or None if\
        \ folder doesn't exist. \n\nRaises:\n    HTTPException (404): If the folder\
        \ is not found."
      operationId: get_files_api_v1_files_files_get
      security:
        - APIKeyQuery: []
        - APIKeyHeader: []
      parameters:
        - name: path
          in: query
          required: true
          schema:
            type: string
            title: Path
        - name: trailarr_api_key
          in: cookie
          required: false
          schema:
            anyOf:
              - type: string
              - type: 'null'
            title: Trailarr Api Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: '#/components/schemas/FolderInfo'
                  - type: 'null'
                title: Response Get Files Api V1 Files Files Get
        '401':
          description: API key is missing
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/media/{media_id}/files:
    get:
      tags:
        - Media
      summary: Get Media Files
      description: "Get media files by ID. \n\nArgs:\n    media_id (int): ID of the\
        \ media item. \n\nReturns:\n    FolderInfo|str: Folder information or error\
        \ message."
      operationId: get_media_files_api_v1_media__media_id__files_get
      security:
        - APIKeyQuery: []
        - APIKeyHeader: []
      parameters:
        - name: media_id
          in: path
          required: true
          schema:
            type: integer
            title: Media Id
        - name: trailarr_api_key
          in: cookie
          required: false
          schema:
            anyOf:
              - type: string
              - type: 'null'
            title: Trailarr Api Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: '#/components/schemas/FolderInfo'
                  - type: string
                title: Response Get Media Files Api V1 Media  Media Id  Files Get
        '401':
          description: API key is missing
        '404':
          description: Media Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/media/{media_id}/trailer:
    delete:
      tags:
        - Media
      summary: Delete Media Trailer
      description: "Delete trailer for media by ID. \n\nArgs:\n    media_id (int):\
        \ ID of the media item. \n\nReturns:\n    str: Deleting trailer message."
      operationId: delete_media_trailer_api_v1_media__media_id__trailer_delete
      security:
        - APIKeyQuery: []
        - APIKeyHeader: []
      parameters:
        - name: media_id
          in: path
          required: true
          schema:
            type: integer
            title: Media Id
        - name: trailarr_api_key
          in: cookie
          required: false
          schema:
            anyOf:
              - type: string
              - type: 'null'
            title: Trailarr Api Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: string
                title: Response Delete Media Trailer Api V1 Media  Media Id  Trailer
                  Delete
        '401':
          description: API key is missing
        '404':
          description: Media Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/trailerprofiles/:
    get:
      tags:
        - Trailer Profiles
      summary: Get all trailer profiles
      description: Get all trailer profiles
      operationId: get_trailer_profiles_api_v1_trailerprofiles__get
      security:
        - APIKeyQuery: []
        - APIKeyHeader: []
      parameters:
        - name: trailarr_api_key
          in: cookie
          required: false
          schema:
            anyOf:
              - type: string
              - type: 'null'
            title: Trailarr Api Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TrailerProfileRead'
                title: Response Get Trailer Profiles Api V1 Trailerprofiles  Get
        '401':
          description: API key is missing
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
        - Trailer Profiles
      summary: Create a new trailer profile
      description: Create a new trailer profile
      operationId: create_trailer_profile_api_v1_trailerprofiles__post
      security:
        - APIKeyQuery: []
        - APIKeyHeader: []
      parameters:
        - name: trailarr_api_key
          in: cookie
          required: false
          schema:
            anyOf:
              - type: string
              - type: 'null'
            title: Trailarr Api Key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TrailerProfileCreate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrailerProfileRead'
        '401':
          description: API key is missing
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/trailerprofiles/{trailerprofile_id}:
    get:
      tags:
        - Trailer Profiles
      summary: Get a trailer profile by ID
      description: Get a trailer profile by ID
      operationId: get_trailer_profile_api_v1_trailerprofiles__trailerprofile_id__get
      security:
        - APIKeyQuery: []
        - APIKeyHeader: []
      parameters:
        - name: trailerprofile_id
          in: path
          required: true
          schema:
            type: integer
            title: Trailerprofile Id
        - name: trailarr_api_key
          in: cookie
          required: false
          schema:
            anyOf:
              - type: string
              - type: 'null'
            title: Trailarr Api Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrailerProfileRead'
        '401':
          description: API key is missing
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
        - Trailer Profiles
      summary: Update a trailer profile
      description: Update a trailer profile
      operationId: update_trailer_profile_api_v1_trailerprofiles__trailerprofile_id__put
      security:
        - APIKeyQuery: []
        - APIKeyHeader: []
      parameters:
        - name: trailerprofile_id
          in: path
          required: true
          schema:
            type: integer
            title: Trailerprofile Id
        - name: trailarr_api_key
          in: cookie
          required: false
          schema:
            anyOf:
              - type: string
              - type: 'null'
            title: Trailarr Api Key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TrailerProfileCreate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrailerProfileRead'
        '401':
          description: API key is missing
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
        - Trailer Profiles
      summary: Delete a trailer profile
      description: Delete a trailer profile
      operationId: delete_trailer_profile_api_v1_trailerprofiles__trailerprofile_id__delete
      security:
        - APIKeyQuery: []
        - APIKeyHeader: []
      parameters:
        - name: trailerprofile_id
          in: path
          required: true
          schema:
            type: integer
            title: Trailerprofile Id
        - name: trailarr_api_key
          in: cookie
          required: false
          schema:
            anyOf:
              - type: string
              - type: 'null'
            title: Trailarr Api Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: boolean
                title: Response Delete Trailer Profile Api V1 Trailerprofiles  Trailerprofile
                  Id  Delete
        '401':
          description: API key is missing
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/trailerprofiles/{trailerprofile_id}/setting:
    post:
      tags:
        - Trailer Profiles
      summary: Update a trailer profile setting
      description: Update a trailer profile setting
      operationId: update_trailer_profile_setting_api_v1_trailerprofiles__trailerprofile_id__setting_post
      security:
        - APIKeyQuery: []
        - APIKeyHeader: []
      parameters:
        - name: trailerprofile_id
          in: path
          required: true
          schema:
            type: integer
            title: Trailerprofile Id
        - name: trailarr_api_key
          in: cookie
          required: false
          schema:
            anyOf:
              - type: string
              - type: 'null'
            title: Trailarr Api Key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSetting'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrailerProfileRead'
        '401':
          description: API key is missing
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
...
    CustomFilterCreate:
      properties:
        filter_name:
          type: string
          title: Filter Name
        filter_type:
          $ref: '#/components/schemas/FilterType'
          default: TRAILER
        id:
          anyOf:
            - type: integer
            - type: 'null'
          title: Id
        filters:
          items:
            $ref: '#/components/schemas/FilterCreate'
          type: array
          title: Filters
          default: []
      type: object
      required:
        - filter_name
      title: CustomFilterCreate
      description: Model for creating/updating CustomFilter.
    CustomFilterRead:
      properties:
        filter_name:
          type: string
          title: Filter Name
        filter_type:
          $ref: '#/components/schemas/FilterType'
          default: TRAILER
        id:
          type: integer
          title: Id
        filters:
          items:
            $ref: '#/components/schemas/FilterRead'
          type: array
          title: Filters
      type: object
      required:
        - filter_name
        - id
        - filters
      title: CustomFilterRead
      description: Model for reading CustomFilter.
    ErrorResponse:
      properties:
        message:
          type: string
          title: Message
      type: object
      required:
        - message
      title: ErrorResponse
    FilterCondition:
      type: string
      enum:
        - EQUALS
        - IS_AFTER
        - IS_BEFORE
        - IN_THE_LAST
        - NOT_IN_THE_LAST
        - NOT_EQUALS
        - GREATER_THAN
        - GREATER_THAN_EQUAL
        - LESS_THAN
        - LESS_THAN_EQUAL
        - CONTAINS
        - NOT_CONTAINS
        - STARTS_WITH
        - NOT_STARTS_WITH
        - ENDS_WITH
        - NOT_ENDS_WITH
        - IS_EMPTY
        - IS_NOT_EMPTY
      title: FilterCondition
    FilterCreate:
      properties:
        filter_by:
          type: string
          title: Filter By
        filter_condition:
          $ref: '#/components/schemas/FilterCondition'
        filter_value:
          type: string
          title: Filter Value
        id:
          anyOf:
            - type: integer
            - type: 'null'
          title: Id
        customfilter_id:
          anyOf:
            - type: integer
            - type: 'null'
          title: Customfilter Id
      type: object
      required:
        - filter_by
        - filter_condition
        - filter_value
      title: FilterCreate
      description: Model for creating/updating Filter.
    FilterRead:
      properties:
        filter_by:
          type: string
          title: Filter By
        filter_condition:
          $ref: '#/components/schemas/FilterCondition'
        filter_value:
          type: string
          title: Filter Value
        id:
          type: integer
          title: Id
        customfilter_id:
          type: integer
          title: Customfilter Id
      type: object
      required:
        - filter_by
        - filter_condition
        - filter_value
        - id
        - customfilter_id
      title: FilterRead
      description: Model for reading Filter.
    FilterType:
      type: string
      enum:
        - HOME
        - MOVIES
        - SERIES
        - TRAILER
      title: FilterType
    FolderInfo:
      properties:
        type:
          type: string
          title: Type
          description: Type of the entry (file/folder).
          default: folder
        name:
          type: string
          title: Name
          description: Name of the file/folder.
        size:
          type: string
          title: Size
          description: Size of the file/folder in human-readable format (e.g. '10
            KB').
          default: 0 KB
        path:
          type: string
          title: Path
          description: Path of the file/folder.
        files:
          items:
            $ref: '#/components/schemas/FolderInfo'
          type: array
          title: Files
          description: List of FileInfo or FolderInfo objects representing files and
            folders         inside a given folder.
          default: []
        created:
          type: string
          title: Created
          description: Creation time of the folder in 'YYYY-MM-DD HH:MM:SS' format.
      type: object
      required:
        - name
        - path
        - created
      title: FolderInfo
      description: "Contains information about files and (sub)folders in a given folder.\
        \ \n\nAttrs:\n    - type (str): Type of the entry (folder).\n    - name (str):\
        \ Name of the folder.\n    - files (list[FileInfo | FolderInfo]): List of\
        \ FileInfo or FolderInfo\n        objects representing files and folders inside\
        \ a given folder.\n    - created (str): Creation time of the folder in \"\
        YYYY-MM-DD HH:MM:SS\"\n        format."
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    TrailerProfileCreate:
      properties:
        enabled:
          type: boolean
          title: Enabled
          default: true
        priority:
          type: integer
          title: Priority
          default: 0
        file_format:
          type: string
          title: File Format
          default: mkv
        file_name:
          type: string
          title: File Name
          default: '{title} ({year})-trailer.{ext}'
        folder_enabled:
          type: boolean
          title: Folder Enabled
          default: false
        folder_name:
          type: string
          title: Folder Name
          default: Trailers
        embed_metadata:
          type: boolean
          title: Embed Metadata
          default: true
        remove_silence:
          type: boolean
          title: Remove Silence
          default: false
        audio_format:
          type: string
          title: Audio Format
          default: opus
        audio_volume_level:
          type: integer
          title: Audio Volume Level
          default: 100
        video_format:
          type: string
          title: Video Format
          default: vp9
        video_resolution:
          type: integer
          title: Video Resolution
          default: 1080
        subtitles_enabled:
          type: boolean
          title: Subtitles Enabled
          default: true
        subtitles_format:
          type: string
          title: Subtitles Format
          default: srt
        subtitles_language:
          type: string
          title: Subtitles Language
          default: en
        search_query:
          type: string
          title: Search Query
          default: '{title} {year} {is_movie} trailer'
        min_duration:
          type: integer
          title: Min Duration
          default: 60
        max_duration:
          type: integer
          title: Max Duration
          default: 600
        always_search:
          type: boolean
          title: Always Search
          default: false
        exclude_words:
          type: string
          title: Exclude Words
          default: ''
        include_words:
          type: string
          title: Include Words
          default: ''
        ytdlp_extra_options:
          type: string
          title: Ytdlp Extra Options
          default: ''
        id:
          anyOf:
            - type: integer
            - type: 'null'
          title: Id
        customfilter_id:
          anyOf:
            - type: integer
            - type: 'null'
          title: Customfilter Id
        customfilter:
          $ref: '#/components/schemas/CustomFilterCreate'
      type: object
      required:
        - customfilter
      title: TrailerProfileCreate
      description: Model for creating/updating TrailerProfile.
    TrailerProfileRead:
      properties:
        enabled:
          type: boolean
          title: Enabled
          default: true
        priority:
          type: integer
          title: Priority
          default: 0
        file_format:
          type: string
          title: File Format
          default: mkv
        file_name:
          type: string
          title: File Name
          default: '{title} ({year})-trailer.{ext}'
        folder_enabled:
          type: boolean
          title: Folder Enabled
          default: false
        folder_name:
          type: string
          title: Folder Name
          default: Trailers
        embed_metadata:
          type: boolean
          title: Embed Metadata
          default: true
        remove_silence:
          type: boolean
          title: Remove Silence
          default: false
        audio_format:
          type: string
          title: Audio Format
          default: opus
        audio_volume_level:
          type: integer
          title: Audio Volume Level
          default: 100
        video_format:
          type: string
          title: Video Format
          default: vp9
        video_resolution:
          type: integer
          title: Video Resolution
          default: 1080
        subtitles_enabled:
          type: boolean
          title: Subtitles Enabled
          default: true
        subtitles_format:
          type: string
          title: Subtitles Format
          default: srt
        subtitles_language:
          type: string
          title: Subtitles Language
          default: en
        search_query:
          type: string
          title: Search Query
          default: '{title} {year} {is_movie} trailer'
        min_duration:
          type: integer
          title: Min Duration
          default: 60
        max_duration:
          type: integer
          title: Max Duration
          default: 600
        always_search:
          type: boolean
          title: Always Search
          default: false
        exclude_words:
          type: string
          title: Exclude Words
          default: ''
        include_words:
          type: string
          title: Include Words
          default: ''
        ytdlp_extra_options:
          type: string
          title: Ytdlp Extra Options
          default: ''
        id:
          type: integer
          title: Id
        customfilter_id:
          type: integer
          title: Customfilter Id
        customfilter:
          $ref: '#/components/schemas/CustomFilterRead'
      type: object
      required:
        - id
        - customfilter_id
        - customfilter
      title: TrailerProfileRead
      description: Model for reading TrailerProfile.
...
    UpdateSetting:
      properties:
        key:
          type: string
          title: Key
        value:
          anyOf:
            - type: integer
            - type: string
            - type: boolean
          title: Value
      type: object
      required:
        - key
        - value
      title: UpdateSetting
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
              - type: string
              - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
        - loc
        - msg
        - type
      title: ValidationError
...
  securitySchemes:
    APIKeyQuery:
      type: apiKey
      in: query
      name: api_key
    APIKeyHeader:
      type: apiKey
      in: header
      name: X-API-KEY